zen_pipeline/pyproject.toml
[project]
authors = [
  {name = "erosons", email = "eromonsei.o.samson@gmail.com"},
]
dependencies = [
  "requests (>=2.32.4,<3.0.0)",
  "boto3 (>=1.39.3,<2.0.0)",
  "pandas (>=2.3.0,<3.0.0)",
  "pendulum (>=3.1.0,<4.0.0)",
  "python-dotenv (>=1.1.1,<2.0.0)"
  ]
description = "to design various pipelinnes patterns"
dynamic = ["version"] # Version is dynamically set in the build process
name = "zen-pipeline"
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry]
exclude = ["event_base_pipeline"] 
packages = [
  {include = "zen_pipeline", from = "utils", format = ["sdist", "wheel"]},
] 
version = "0.1.0" # base version, will be overridden by dynamic versioning

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[project.urls]
documentation = "https://github.com/myDatapipeline/Engineering_Piplines_Guide/blob/main/README.md"
homepage = "https://github.com/myDatapipeline/Engineering_Piplines_Guide"
issue-tracker = "https://github.com/myDatapipeline/Engineering_Piplines_Guide/issues"
repository = "https://github.com/myDatapipeline/Engineering_Piplines_Guide"

# [[tool.poetry.source]]
# name = "private"
# url = "https://pypi.org/simple"
